backend/pyproject.toml
[project]
name = "bus-reservation"
version = "0.1.0"
description = "Bus Reservation System with Seat Locking"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "qrcode[pil]>=7.4.2",
    "apscheduler>=3.10.4",
    "email-validator>=2.1.0",
]
